+++ benchmarks/StarWarsBench.php
@@ -1,4 +1,7 @@
 <?php
+
+declare(strict_types=1);
+
 namespace GraphQL\Benchmarks;
 
 use GraphQL\GraphQL;
@@ -14,19 +17,19 @@ use GraphQL\Type\Introspection;
  */
 class StarWarsBench
 {
-    private $introQuery;
+    private string $introQuery;
 
-    public function setIntroQuery()
+    public function setIntroQuery(): void
     {
         $this->introQuery = Introspection::getIntrospectionQuery();
     }
 
-    public function benchSchema()
+    public function benchSchema(): void
     {
         StarWarsSchema::build();
     }
 
-    public function benchHeroQuery()
+    public function benchHeroQuery(): void
     {
         $q = '
         query HeroNameQuery {
@@ -42,7 +45,7 @@ class StarWarsBench
         );
     }
 
-    public function benchNestedQuery()
+    public function benchNestedQuery(): void
     {
         $q = '
         query NestedQuery {
@@ -64,7 +67,7 @@ class StarWarsBench
         );
     }
 
-    public function benchQueryWithFragment()
+    public function benchQueryWithFragment(): void
     {
         $q = '
         query UseFragment {
@@ -88,7 +91,7 @@ class StarWarsBench
         );
     }
 
-    public function benchStarWarsIntrospectionQuery()
+    public function benchStarWarsIntrospectionQuery(): void
     {
         GraphQL::executeQuery(
             StarWarsSchema::build(),