+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Removed deprecated `Type::getInternalTypes()`
 - Remove deprecated `GraphQL::execute()`
 - Remove deprecated `GraphQL::executeAndReturnResult()`
+- Removed depreacted experimental CoroutineExecutor
 
 #### 14.6.1