+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
 #### Unreleased
+Changed:
 - PHP version required: 7.4+
+
+Removed:
 - Removed deprecated `Type::getInternalTypes()`
 - Remove deprecated `GraphQL::execute()`
 - Remove deprecated `GraphQL::executeAndReturnResult()`
+
+Fix:
+- Fix overly eager validation of repeatable directive usage
 
 #### 14.6.1